Move commands to respective folders
