+++ src/main/java/frc/robot/commands/BallIntake/DefaultIntake.java
@@ -5,33 +5,37 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.BallIntake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.OI;
 import frc.robot.TestingDashboard;
 import frc.robot.input.XboxController.XboxAxis;
-import frc.robot.subsystems.Turret;
+import frc.robot.subsystems.BallIntake;
 
-public class DefaultTurret extends CommandBase {
-  /**
-   * Creates a new DefaultTurret.
-   */
-  Turret m_turret;
+/**
+ * A Default Intake command that passes XBox controller input to the Ball Intake.
+ */
+public class DefaultIntake extends CommandBase {
+  @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
+  private final BallIntake m_intake;
   private static OI oi;
 
-  public DefaultTurret() {
-    // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(Turret.getInstance());
+  /**
+   * Creates a new DefaultIntake.
+   * @param BallIntake the subsystem used by this command.
+   */
+  public DefaultIntake(BallIntake intake) {
+    m_intake = intake;
     oi = OI.getInstance();
-    m_turret = Turret.getInstance();
+    // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(m_intake);
   }
 
   public static void registerWithTestingDashboard() {
-    Turret turret = Turret.getInstance();
-    DefaultTurret cmd = new DefaultTurret();
-    TestingDashboard.getInstance().registerCommand(turret, "Basic", cmd);
-
+    BallIntake intake = BallIntake.getInstance();
+    DefaultIntake cmd = new DefaultIntake(BallIntake.getInstance());
+    TestingDashboard.getInstance().registerCommand(intake, "Basic", cmd);
   }
 
   // Called when the command is initially scheduled.
@@ -42,9 +46,9 @@ public class DefaultTurret extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    // Drives the Turret with the x-axis of the left Xbox joystick.
-    double speed = oi.getXbox().getAxis(XboxAxis.kXRight);
-    m_turret.spinTurretMotor(0.75*speed);
+    //Drives the ball intake with the y-axis of the left XBox joystick
+    double speed = oi.getXbox().getAxis(XboxAxis.kRightTrigger);
+    m_intake.spinIntakeRoller(speed);
   }
 
   // Called once the command ends or is interrupted.
@@ -53,6 +57,7 @@ public class DefaultTurret extends CommandBase {
   }
 
   // Returns true when the command should end.
+  //Default command so will never finish running
   @Override
   public boolean isFinished() {
     return false;
