+++ src/main/java/frc/robot/commands/BallIntake/SpinIntakeRoller.java
@@ -5,37 +5,32 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+// Spins the ball intake roller while command is active.
+package frc.robot.commands.BallIntake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.OI;
 import frc.robot.TestingDashboard;
-import frc.robot.input.XboxController.XboxAxis;
 import frc.robot.subsystems.BallIntake;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
-/**
- * A Default Intake command that passes XBox controller input to the Ball Intake.
- */
-public class DefaultIntake extends CommandBase {
-  @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-  private final BallIntake m_intake;
-  private static OI oi;
-
+public class SpinIntakeRoller extends CommandBase {
   /**
-   * Creates a new DefaultIntake.
-   * @param BallIntake the subsystem used by this command.
+   * Creates a new SpinIntakeRoller.
    */
-  public DefaultIntake(BallIntake intake) {
-    m_intake = intake;
-    oi = OI.getInstance();
+
+   BallIntake m_ballIntake;
+
+  public SpinIntakeRoller() {
     // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(m_intake);
+
+    addRequirements(BallIntake.getInstance());
+    m_ballIntake = BallIntake.getInstance();
   }
 
   public static void registerWithTestingDashboard() {
-    BallIntake intake = BallIntake.getInstance();
-    DefaultIntake cmd = new DefaultIntake(BallIntake.getInstance());
-    TestingDashboard.getInstance().registerCommand(intake, "Basic", cmd);
+    BallIntake ballIntake = BallIntake.getInstance();
+    SpinIntakeRoller cmd = new SpinIntakeRoller();
+    TestingDashboard.getInstance().registerCommand(ballIntake, "Basic", cmd);
   }
 
   // Called when the command is initially scheduled.
@@ -46,18 +41,19 @@ public class DefaultIntake extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    //Drives the ball intake with the y-axis of the left XBox joystick
-    double speed = oi.getXbox().getAxis(XboxAxis.kRightTrigger);
-    m_intake.spinIntakeRoller(speed);
+
+    double speed = SmartDashboard.getNumber("IntakeRollerSpeed",0.5);
+    m_ballIntake.spinIntakeRoller(speed);
+
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
+    m_ballIntake.spinIntakeRoller(0);
   }
 
   // Returns true when the command should end.
-  //Default command so will never finish running
   @Override
   public boolean isFinished() {
     return false;
