+++ src/main/java/frc/robot/commands/Conveyor/PrepareBallsToShoot.java
@@ -5,56 +5,55 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-// Spins the conveyor motor for a specified amount of time.
-package frc.robot.commands;
+// Spins the conveyor motor until there is a ball ready to shoot.
+package frc.robot.commands.Conveyor;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.TestingDashboard;
-import frc.robot.subsystems.BallIntake;
 import frc.robot.subsystems.Conveyor;
+import frc.robot.TestingDashboard;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
-public class SpinConveyorHTimed extends CommandBase {
+public class PrepareBallsToShoot extends CommandBase {
   /**
-   * Creates a new SpinConveyor1Timed.
+   * Creates a new PrepareBallsToShoot.
    */
 
-   Timer m_timer;
    Conveyor m_conveyor;
-   double m_period;
+   Timer m_timer;
+   private static final int m_period = 10;
 
-  public SpinConveyorHTimed() {
+  public PrepareBallsToShoot() {
     // Use addRequirements() here to declare subsystem dependencies.
-    
-    addRequirements(Conveyor.getInstance());
-    m_timer = new Timer();
-    m_conveyor = Conveyor.getInstance();
 
+    addRequirements(Conveyor.getInstance());
+    m_conveyor = Conveyor.getInstance();
+    m_timer = new Timer();
+    
   }
 
   public static void registerWithTestingDashboard() {
     Conveyor conveyor = Conveyor.getInstance();
-    SpinConveyorHTimed cmd = new SpinConveyorHTimed();
-    TestingDashboard.getInstance().registerCommand(conveyor, "Timed", cmd);
+    PrepareBallsToShoot cmd = new PrepareBallsToShoot();
+    TestingDashboard.getInstance().registerCommand(conveyor, "Basic", cmd);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    m_period = SmartDashboard.getNumber("ConveyorHMotoryTimeout", 5); // default of 5 seconds
     m_timer.start();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-
-    double speed1 = SmartDashboard.getNumber("ConveyorHMotor1Speed",0.5);
-    double speed2 = SmartDashboard.getNumber("ConveyorHMotor2Speed",0.5);
-    m_conveyor.spinHConveyor1(speed1);
-    m_conveyor.spinHConveyor2(speed2);
-
+    double speed = SmartDashboard.getNumber("Conveyor1MotorSpeed",0.5);
+    if (m_conveyor.ballReadyToShoot() == false){
+      m_conveyor.spinHConveyors(speed);
+    }
+    if (m_conveyor.ballReadyToShoot() == true){
+      m_conveyor.spinHConveyors(0);
+    }
   }
 
   // Called once the command ends or is interrupted.
@@ -67,6 +66,6 @@ public class SpinConveyorHTimed extends CommandBase {
   @Override
   public boolean isFinished() {
     boolean timerExpired = m_timer.hasPeriodPassed(m_period);
-    return timerExpired;
+    return (timerExpired || (m_conveyor.ballReadyToShoot() == true));
   }
 }
