+++ src/main/java/frc/robot/commands/Conveyor/SpinBothConveyorsTimed.java
@@ -6,11 +6,10 @@
 /*----------------------------------------------------------------------------*/
 
 // Spins the conveyor motor for a specified amount of time.
-package frc.robot.commands;
+package frc.robot.commands.Conveyor;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.TestingDashboard;
-import frc.robot.subsystems.BallIntake;
 import frc.robot.subsystems.Conveyor;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
