+++ src/main/java/frc/robot/commands/Shooter/SpinShooter.java
@@ -5,46 +5,50 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.Shooter;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.TestingDashboard;
-import frc.robot.subsystems.Turret;
+import frc.robot.subsystems.Shooter;
 
-public class TurretLeft extends CommandBase {
-  Turret m_turret;
+public class SpinShooter extends CommandBase {
   /**
-   * Creates a new TurretLeft.
+   * Creates a new SpinShooter.
    */
-  public TurretLeft() {
+  Shooter m_shooter;
+
+  public SpinShooter() {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_turret = Turret.getInstance();
-    addRequirements(m_turret);
+    addRequirements(Shooter.getInstance());
+    m_shooter = Shooter.getInstance();
   }
 
   public static void registerWithTestingDashboard() {
-    Turret turret = Turret.getInstance();
-    TurretLeft cmd = new TurretLeft();
-    TestingDashboard.getInstance().registerCommand(turret, "Basic", cmd);
+    Shooter shooter = Shooter.getInstance();
+    SpinShooter cmd = new SpinShooter();
+    TestingDashboard.getInstance().registerCommand(shooter, "Basic", cmd);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    double speed = SmartDashboard.getNumber("IncrementTurretMotorSpeed", 0.5);
-    m_turret.spinTurretMotor(-speed);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
+    double topSpeed = SmartDashboard.getNumber("TopShooterSpeed",0);
+    double botSpeed = SmartDashboard.getNumber("BottomShooterSpeed",0);
+    m_shooter.setTop(-topSpeed);
+    m_shooter.setBottom(-botSpeed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_turret.spinTurretMotor(0);
+    m_shooter.setTop(0);
+    m_shooter.setBottom(0);
   }
 
   // Returns true when the command should end.
