+++ src/main/java/frc/robot/commands/Spinner/DeploySpinner.java
@@ -7,7 +7,7 @@
 
 //This command extends the piston that is used to push up the spinner motor and color sensor.
 
-package frc.robot.commands;
+package frc.robot.commands.Spinner;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;  
 import frc.robot.subsystems.Spinner;
