+++ src/main/java/frc/robot/commands/Spinner/SpinSpinnerMotorTimed.java
@@ -5,55 +5,60 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+//This command spins the spinner motor for a certain amount of time.
 
+package frc.robot.commands.Spinner;
+
+import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.TestingDashboard;
-import frc.robot.subsystems.Shooter;
+import frc.robot.subsystems.Spinner;
 
-public class SpinShooter extends CommandBase {
+public class SpinSpinnerMotorTimed extends CommandBase {
+  Timer m_timer;
+  double m_period;
+  Spinner m_spinner;
   /**
-   * Creates a new SpinShooter.
+   * Creates a new SpinSpinnerMotorTimed.
    */
-  Shooter m_shooter;
-
-  public SpinShooter() {
+  public SpinSpinnerMotorTimed() {
     // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(Shooter.getInstance());
-    m_shooter = Shooter.getInstance();
+    addRequirements(Spinner.getInstance());
+    m_timer = new Timer();
+    m_spinner = Spinner.getInstance();
   }
 
   public static void registerWithTestingDashboard() {
-    Shooter shooter = Shooter.getInstance();
-    SpinShooter cmd = new SpinShooter();
-    TestingDashboard.getInstance().registerCommand(shooter, "Basic", cmd);
+    Spinner spinner = Spinner.getInstance();
+    SpinSpinnerMotorTimed cmd = new SpinSpinnerMotorTimed();
+    TestingDashboard.getInstance().registerCommand(spinner, "Timed", cmd);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
+    m_timer.start();
+    m_period = SmartDashboard.getNumber("SpinnerMotorPeriod", 5);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    double topSpeed = SmartDashboard.getNumber("TopShooterSpeed",0);
-    double botSpeed = SmartDashboard.getNumber("BottomShooterSpeed",0);
-    m_shooter.setTop(-topSpeed);
-    m_shooter.setBottom(-botSpeed);
+    double speed = SmartDashboard.getNumber("SpinnerMotorSpeed",0.2);
+    m_spinner.spin(speed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_shooter.setTop(0);
-    m_shooter.setBottom(0);
+    m_spinner.spin(0);
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    boolean timerExpired = m_timer.hasPeriodPassed(m_period);
+    return timerExpired;
   }
 }
