+++ src/main/java/frc/robot/commands/Turret/DefaultTurret.java
@@ -5,44 +5,51 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.Turret;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.OI;
 import frc.robot.TestingDashboard;
+import frc.robot.input.XboxController.XboxAxis;
 import frc.robot.subsystems.Turret;
 
-public class TurretStop extends CommandBase {
-  Turret m_turret;
+public class DefaultTurret extends CommandBase {
   /**
-   * Creates a new TurretStop.
+   * Creates a new DefaultTurret.
    */
-  public TurretStop() {
+  Turret m_turret;
+  private static OI oi;
+
+  public DefaultTurret() {
     // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(Turret.getInstance());
+    oi = OI.getInstance();
     m_turret = Turret.getInstance();
-    addRequirements(m_turret);
   }
 
   public static void registerWithTestingDashboard() {
     Turret turret = Turret.getInstance();
-    TurretStop cmd = new TurretStop();
+    DefaultTurret cmd = new DefaultTurret();
     TestingDashboard.getInstance().registerCommand(turret, "Basic", cmd);
+
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    m_turret.spinTurretMotor(0);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
+    // Drives the Turret with the x-axis of the left Xbox joystick.
+    double speed = oi.getXbox().getAxis(XboxAxis.kXRight);
+    m_turret.spinTurretMotor(0.5*speed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_turret.spinTurretMotor(0);
   }
 
   // Returns true when the command should end.
