+++ src/main/java/frc/robot/commands/Turret/TurretRight.java
@@ -1,0 +1,55 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package frc.robot.commands.Turret;
+
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.TestingDashboard;
+import frc.robot.subsystems.Turret;
+
+public class TurretRight extends CommandBase {
+  Turret m_turret;
+  /**
+   * Creates a new TurretRight.
+   */
+  public TurretRight() {
+    // Use addRequirements() here to declare subsystem dependencies.
+    m_turret = Turret.getInstance();
+    addRequirements(m_turret);
+  }
+
+  public static void registerWithTestingDashboard() {
+    Turret turret = Turret.getInstance();
+    TurretRight cmd = new TurretRight();
+    TestingDashboard.getInstance().registerCommand(turret, "Basic", cmd);
+  }
+
+  // Called when the command is initially scheduled.
+  @Override
+  public void initialize() {
+    double speed = SmartDashboard.getNumber("IncrementTurretMotorSpeed", 0.5);
+    m_turret.spinTurretMotor(speed);
+  }
+
+  // Called every time the scheduler runs while the command is scheduled.
+  @Override
+  public void execute() {
+  }
+
+  // Called once the command ends or is interrupted.
+  @Override
+  public void end(boolean interrupted) {
+    m_turret.spinTurretMotor(0);
+  }
+
+  // Returns true when the command should end.
+  @Override
+  public boolean isFinished() {
+    return false;
+  }
+}
