+++ src/main/java/frc/robot/commands/Turret/TurretStop.java
@@ -5,59 +5,49 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-//This command retracts the piston that is used to push up the spinner motor and color sensor.
-
-package frc.robot.commands;
+package frc.robot.commands.Turret;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import edu.wpi.first.wpilibj.DoubleSolenoid;
-import frc.robot.subsystems.Spinner;
-
 import frc.robot.TestingDashboard;
+import frc.robot.subsystems.Turret;
 
-public class RetractSpinner extends CommandBase {
-  Spinner m_spinner;
-  boolean m_finished = false;
-  DoubleSolenoid m_piston;
-  
-  public RetractSpinner() {
+public class TurretStop extends CommandBase {
+  Turret m_turret;
+  /**
+   * Creates a new TurretStop.
+   */
+  public TurretStop() {
     // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(Spinner.getInstance());
-    m_spinner = Spinner.getInstance();
-    m_piston = m_spinner.getPiston();
+    m_turret = Turret.getInstance();
+    addRequirements(m_turret);
   }
 
   public static void registerWithTestingDashboard() {
-    Spinner spinner = Spinner.getInstance();
-    RetractSpinner cmd = new RetractSpinner();
-    TestingDashboard.getInstance().registerCommand(spinner, "Basic", cmd);
+    Turret turret = Turret.getInstance();
+    TurretStop cmd = new TurretStop();
+    TestingDashboard.getInstance().registerCommand(turret, "Basic", cmd);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    m_piston.set(DoubleSolenoid.Value.kReverse);
-    m_finished = false;
+    m_turret.spinTurretMotor(0);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    if (m_piston.get() == DoubleSolenoid.Value.kReverse) {
-      m_piston.set(DoubleSolenoid.Value.kOff);
-      m_finished = true;
-    }
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_piston.set(DoubleSolenoid.Value.kOff); 
+    m_turret.spinTurretMotor(0);
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return m_finished;
+    return false;
   }
 }
